TokenAuthentication class and unit tests
