+++ src/Services/Services.csproj
@@ -332,6 +332,7 @@
     <Compile Include="Virtualization\SenseNetVideoStreamerHandler.cs" />
     <Compile Include="Virtualization\StaticVirtualizedFileHandler.cs" />
     <Compile Include="Virtualization\StaticVirtualizedFileHandlerFactory.cs" />
+    <Compile Include="Virtualization\TokenAuthentication.cs" />
     <Compile Include="WebDAV\Common.cs" />
     <Compile Include="WebDAV\Delete.cs" />
     <Compile Include="WebDAV\Enums.cs" />
